+++ ssr/src/component/mod.rs
@@ -1,34 +1,9 @@
-pub mod airdrop_logo;
-pub mod auth_providers;
-pub mod back_btn;
-pub mod base_route;
-pub mod bullet_loader;
-pub mod buttons;
-pub mod canisters_prov;
-pub mod coming_soon;
-pub mod connect;
-pub mod content_upload;
-pub mod dashbox;
-pub mod feed_popup;
-pub mod hn_icons;
-pub mod ic_symbol;
-pub mod icons;
-pub mod infinite_scroller;
-pub mod loading;
-pub mod login_modal;
-pub mod modal;
-pub mod nav;
-pub mod nav_icons;
-pub mod onboarding_flow;
-pub mod option;
-pub mod overlay;
-pub mod profile_placeholders;
-pub mod scrolling_post_view;
-pub mod share_popup;
-pub mod social;
-pub mod spinner;
-pub mod title;
-pub mod toggle;
-pub mod token_confetti_symbol;
-pub mod token_logo_sanitize;
-pub mod video_player;
+pub mod icpump;
+pub mod yral; 
+pub mod shared;
+
+#[derive(Clone, Copy, PartialEq)] 
+pub enum AppType {
+    Yral,
+    IcPump,
+}
