+++ ssr/src/component/shared/assets/airdrop_logo.rs
