+++ ssr/src/component/shared/assets/back_btn.rs
