+++ ssr/src/component/shared/assets/bullet_loader.rs
