+++ ssr/src/component/shared/assets/buttons.rs
