+++ ssr/src/component/shared/assets/coming_soon.rs
