+++ ssr/src/component/shared/assets/hn_icons.rs
