+++ ssr/src/component/shared/assets/ic_symbol.rs
