+++ ssr/src/component/shared/assets/icons/airdrop_icon.rs
