+++ ssr/src/component/shared/assets/icons/arrow_left_right_icon.rs
