+++ ssr/src/component/shared/assets/icons/chevron_right_icon.rs
