+++ ssr/src/component/shared/assets/icons/eye_hide_icon.rs
