+++ ssr/src/component/shared/assets/icons/mod.rs
