+++ ssr/src/component/shared/assets/icons/notification_icon.rs
