+++ ssr/src/component/shared/assets/icons/send_icon.rs
