+++ ssr/src/component/shared/assets/icons/share_icon.rs
