+++ ssr/src/component/shared/assets/loading.rs
