+++ ssr/src/component/shared/assets/mod.rs
@@ -1,0 +1,18 @@
+pub mod airdrop_logo;
+pub mod back_btn;
+pub mod bullet_loader;
+pub mod buttons;
+pub mod coming_soon;
+pub mod hn_icons;
+pub mod ic_symbol;
+pub mod icons;
+pub mod loading;
+pub mod modal;
+pub mod nav_icons;
+pub mod onboarding_flow;
+pub mod option;
+pub mod overlay;
+pub mod profile_placeholders;
+pub mod spinner;
+pub mod toggle;
+pub mod token_confetti_symbol;
