+++ ssr/src/component/shared/assets/nav_icons.rs
