+++ ssr/src/component/shared/assets/onboarding_flow.rs
