+++ ssr/src/component/shared/assets/option.rs
