+++ ssr/src/component/shared/assets/overlay.rs
