+++ ssr/src/component/shared/assets/profile_placeholders.rs
