+++ ssr/src/component/shared/assets/spinner.rs
