+++ ssr/src/component/shared/assets/toggle.rs
