+++ ssr/src/component/shared/assets/token_confetti_symbol.rs
