+++ ssr/src/component/shared/components/mod.rs
@@ -1,0 +1,16 @@
+pub mod auth_providers;
+pub mod base_route;
+pub mod canisters_prov;
+pub mod connect;
+pub mod content_upload;
+pub mod dashbox;
+pub mod feed_popup;
+pub mod infinite_scroller;
+pub mod login_modal;
+pub mod nav;
+pub mod scrolling_post_view;
+pub mod share_popup;
+pub mod social;
+pub mod title;
+pub mod token_logo_sanitize;
+pub mod video_player;
