+++ ssr/src/component/shared/components/video_player.rs
