+++ ssr/src/component/shared/mod.rs
@@ -1,0 +1,2 @@
+pub mod assets;
+pub mod components;
