+++ ssr/src/page/view_profile_redirect.rs
@@ -1,4 +1,5 @@
-use crate::component::{canisters_prov::AuthCansProvider, spinner::FullScreenSpinner};
+use crate::component::shared::assets::spinner::FullScreenSpinner;
+use crate::component::shared::components::canisters_prov::AuthCansProvider;
 use leptos::*;
 use leptos_router::Redirect;
 #[component]
